Update README.md
My thesis project was a pilot study investigating the impact of songwriting education and practice on levels of depression, fatigue, anxiety, stress, and life satisfaction in adults diagnosed with Celiac Disease. The study was expanded to include all chronic illness for recruitment purposes. For my dissertation, I would like to conduct a feasibility study to evaluate the validity and adaptability of this intervention on people coping with chronic illness, as well as documenting data that I was not able to collect in the pilot. Ideally, I would also be able to test hypotheses about the cognitive mechanisms and communication theories I’ve used to explain the pilot data.